version BLN_PFE-DRV_S32G_A53_LNX_BETA_0.9.6 RC1

+++ sw/pfe_platform/public/pfe_fp.h
@@ -12,8 +12,8 @@
 #include "pfe_class.h"
 
 void pfe_fp_init(void);
-uint32_t pfe_fp_create_table(pfe_class_t *class, uint8_t rules_count);
-uint32_t pfe_fp_table_write_rule(pfe_class_t *class, uint32_t table_address, const pfe_ct_fp_rule_t *rule, uint8_t position);
+uint32_t pfe_fp_create_table(pfe_class_t *class, uint16_t rules_count);
+uint32_t pfe_fp_table_write_rule(pfe_class_t *class, uint32_t table_address, const pfe_ct_fp_rule_t *rule, uint16_t position);
 void pfe_fp_destroy_table(const pfe_class_t *class, uint32_t table_address);
 errno_t pfe_fp_table_get_statistics(pfe_class_t *class, uint32_t pe_idx ,uint32_t table_address, pfe_ct_class_flexi_parser_stats_t *stats);
